% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_summary.R
\name{boot_summary}
\alias{boot_summary}
\title{Summarising Regression Models Using the Bootstrap}
\usage{
boot_summary(
  model,
  type = "perc",
  method = "residual",
  conf.level = 0.95,
  R = 999,
  pval_precision = NULL,
  adjust.method = "none",
  ...
)
}
\arguments{
\item{model}{An object of class "lm", "glm", or "nls".}

\item{type}{A vector of character strings representing the type of interval to base the test on. The value should be one of "norm", "basic", "stud", "perc" (the default), and "bca".}

\item{method}{Either "residual" (for resampling of scaled and centred residuals, the default) or "case" (for case resampling).}

\item{conf.level}{The confidence level for the confidence intervals. The default is 0.95.}

\item{R}{The number of bootstrap replicates. The default is 999.}

\item{pval_precision}{The desired precision for the p-value. The default is 1/R.}

\item{adjust.method}{Adjustment of p-values for multiple comparisons using \code{p.adjust}. The default is "none", in which case the p-value aren't adjusted. The other options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr"; see \code{?p.adjust} for details on these methods.}

\item{...}{Additional arguments passed to \code{boot}.}
}
\value{
A data frame containing coefficient estimates, bootstrap confidence intervals, and bootstrap p-values.
}
\description{
Summaries for "lm", "glm", and "nls" objects, using the bootstrap for p-values and confidence intervals.
}
\examples{
# Bootstrap summary of a linear model for mtcars:
model <- lm(mpg ~ hp + wt, data = mtcars)
boot_summary(model)

# Adjust p-values for multiplicity using Holm's method:
boot_summary(model, adjust.method = "holm")
}
